requirements.txt
# Core
openai>=1.54.0
streamlit>=1.35.0
plotly>=5.18.0
pandas>=2.2.3
pyarrow>=14.0.0
requests>=2.32.3
python-dotenv>=1.0.1

# Testing
pytest>=8.3.3
